+++ lib/src/clojure_lsp/feature/paredit.clj
@@ -8,7 +8,18 @@
 (defn ^:private paredit-op [paredit-fn move-cursor? uri original-zloc row col]
   (try
     (when original-zloc
-      (let [zloc (paredit-fn original-zloc)
+      (let [{original-row :row original-col :col} (-> original-zloc z/node meta)
+            offset-row (- row original-row)
+            offset-col (- col original-col)
+            pos-zloc (-> original-zloc
+                         z/root-string
+                         (z/of-string {:track-position? true}) ;; https://clojurians.slack.com/archives/CHB5Q2XUJ/p1740958982303849
+                         (edit/find-at-pos row col))
+            zloc (paredit-fn pos-zloc)
+            [row' col'] (z/position zloc)
+            [new-row new-col] (if (= (z/node original-zloc) (z/node zloc)) ;; if the node where the cursor is has changed
+                                [(+ row' offset-row) (+ col' offset-col)]  ;; move the cursor to the new node position
+                                [row col])                                 ;; otherwise keep the cursor in the same place
             root-zloc (z/up (edit/to-top zloc))]
         {:changes-by-uri {uri [{:loc root-zloc
                                 :range (meta (z/node root-zloc))}]} ;; FIXME: range is always the whole document
@@ -16,7 +27,10 @@
                                     :take-focus true
                                     :range (if move-cursor?
                                              (meta (z/node (z/up original-zloc)))
-                                             {:row row :col col :end-row row :end-col col})}}))
+                                             {:row new-row
+                                              :col new-col
+                                              :end-row new-row
+                                              :end-col new-col})}}))
     (catch Exception e
       (logger/error e))))