+++ CHANGELOG.md
@@ -6,7 +6,7 @@
   - Bump clj-kondo to `2025.02.20`.
   - Add support for OpenTelemetry(otlp) log, enabled if configured. #1963
 - Editor
-  - Improve paredit slurp corner cases. #1973
+  - Improve paredit slurp and barf corner cases. #1973 #1976
 
 ## 2025.02.07-16.11.24